Add kstore-file module
