+++ kstore-file/src/appleMain/kotlin/io/github/xxfast/kstore/file/utils/FileSystem.kt
@@ -1,0 +1,5 @@
+package io.github.xxfast.kstore.file.utils
+
+import okio.FileSystem
+
+public actual val FILE_SYSTEM: FileSystem = FileSystem.SYSTEM
