+++ kstore-file/src/commonMain/kotlin/io/github/xxfast/kstore/file/utils/FileSystem.kt
@@ -1,0 +1,8 @@
+package io.github.xxfast.kstore.file.utils
+
+import okio.FileSystem
+
+/***
+ * Okio file system for the given platform
+ */
+public expect val FILE_SYSTEM: FileSystem
