+++ kstore-file/src/jsMain/kotlin/io/github/xxfast/kstore/file/utils/FileSystem.kt
@@ -1,0 +1,6 @@
+package io.github.xxfast.kstore.file.utils
+
+import okio.FileSystem
+import okio.NodeJsFileSystem
+
+public actual val FILE_SYSTEM: FileSystem = NodeJsFileSystem
