--- kstore/src/appleMain/kotlin/io/github/xxfast/kstore/utils/FileSystem.kt
@@ -1,5 +1,0 @@
-package io.github.xxfast.kstore.utils
-
-import okio.FileSystem
-
-public actual val FILE_SYSTEM: FileSystem = FileSystem.SYSTEM
