--- kstore/src/commonMain/kotlin/io/github/xxfast/kstore/utils/FileSystem.kt
@@ -1,8 +1,0 @@
-package io.github.xxfast.kstore.utils
-
-import okio.FileSystem
-
-/***
- * Okio file system for the given platform
- */
-public expect val FILE_SYSTEM: FileSystem
