--- kstore/src/jsMain/kotlin/io/github/xxfast/kstore/utils/FileSystem.kt
@@ -1,6 +1,0 @@
-package io.github.xxfast.kstore.utils
-
-import okio.FileSystem
-import okio.NodeJsFileSystem
-
-public actual val FILE_SYSTEM: FileSystem = NodeJsFileSystem
